pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flashpay"
version = "0.1.0"
description = "Atomic swap+pay on Tempo. Convert and settle in one transaction."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pympp[tempo]>=0.5",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["flashpay"]
